Fixed verb, inheritance, and NPC v NPC
1) Fixed verbMatrix calculations. Java uses % as remainder (positive OR
negative) while other languages use % for modulus (always positive). I
just added code so that if the remainder is negative, I added the
dividend to the remainder to make it a modulus operation.

2) Made Player a interface instead of a class, then PC and NPC classes
inherit the way I had intended. However, also needed to cast the objects
from Player to either (PC) or (NPC) when passing the objects between
classes.

3) Stupid copy / paste bug made p1 and p2 reference the SAME object.
THAT IS WHY they always tied. When I pasted p1's initialization code (PC
or NPC) into p2's initialization code, I only changed ONE reference to
p1 to the correct p2 reference, but not the other reference to p1.

4) Removed debug outputs (since now everything works) and prettied up
the output real nice-like.

************
NEXT COMMIT:
************

1) Final polishing before submitting this Final Project for CSC 222-1H0F
SP23 at Laurel Ridge Community College, and then DEMONSTRATING for the
class.

2) MAY change inheritance structures to add an abstract class between
the Player interface and the PC and NPC classes.